Move testCreateFactoryImmutable and testCreateFactory to new FactoryTest.
